Put DynamoDB autoscaling behind an interface
Move old-style AWS autoscaling to a separate file; make metrics
autoscaling implement the common interface.
