+++ pkg/chunk/aws/metrics_autoscaling_test.go
@@ -19,7 +19,7 @@ func TestTableManagerMetricsAutoScaling(t *testing.T) {
 
 	client := dynamoTableClient{
 		DynamoDB: dynamoDB,
-		metrics: &metricsData{
+		autoscale: &metricsData{
 			promAPI:           &mockProm,
 			queueLengthTarget: 100000,
 			scaleUpFactor:     1.2,
